apps/backend/requirements.txt
# apps/backend/requirements.txt - Clean minimal version
fastapi==0.115.14
uvicorn==0.35.0
openai==1.93.0
python-dotenv==1.1.1
pydantic==2.11.7

# RAG Dependencies
PyPDF2==3.0.1
python-multipart==0.0.6
numpy
qdrant-client